add unit and line number to mcx_error, use -O3, fix all warnings
git-svn-id: https://orbit.nmr.mgh.harvard.edu/svn/mcxcl/trunk@47 4848f366-3658-0410-a320-90152d9d5b07

+++ src/tictoc.c
@@ -74,11 +74,11 @@ static unsigned int timerRes;
 #include <string.h>
 void SetupMillisTimer(void) {}
 void CleanupMillisTimer(void) {}
-int GetTime (void) {
+long GetTime (void) {
   struct timeval tv;
   timerRes = 1000;
   gettimeofday(&tv,NULL);
-  int temp = tv.tv_usec;
+  long temp = tv.tv_usec;
   temp+=tv.tv_sec*1000000;
   return temp;
 }
@@ -99,7 +99,7 @@ unsigned int StartTimer () {
  *      based on the performance counters.
  */
 
-int GetTime(void)
+long GetTime(void)
 {
    static double cycles_per_usec;
    LARGE_INTEGER counter;
@@ -119,7 +119,7 @@ int GetTime(void)
       return 0;
    }
 
-   return ((int) (((double) counter.QuadPart) * cycles_per_usec));
+   return ((long) (((double) counter.QuadPart) * cycles_per_usec));
 }
 
 #pragma comment(lib,"winmm.lib")
